requirements.txt
djangorestframework=3.15.1
django-filter=24.2
django-cors-headers=4.3.1
